+++ src/App.js
@@ -1,0 +1,11 @@
+import './App.css';
+
+function App() {
+  return (
+    <div className="App">
+      <h1>SkyRocket</h1>
+    </div>
+  );
+}
+
+export default App;
